react : admin jsx 파일들 main폴더로 이동, module.css삭제 이후 전체 css 생성 습도 도넛 차트에 가상 데이터를 가지고 오기 전 까지 진행
