+++ React/farm_animals_shop/src/admin_component/main/AdminHeader.jsx
@@ -1,11 +1,10 @@
 import React from 'react'
-import styles from './Admin.module.css'
 
 const AdminHeader = () => {
   // 관리자 상단 header
   return (
     <>
-      <div className={styles.admin_header}>
+      <div>
           <h1>축산축산</h1>
       </div>
     </>
